Encapsulate configuration menu item
